For #154: Refactored the very procedural logic for overwriting the target
