+++ src/main/java/org/jpeek/Target.java
@@ -1,0 +1,9 @@
+package org.jpeek;
+
+import java.io.IOException;
+import java.nio.file.Path;
+
+public interface Target {
+
+	public Path toPath() throws IOException;
+}
